refactor(frr): extract protocol rendering to sibling file
Move the routing-protocol renderer generateProtocols out of
policy_render.go into a cohesive sibling protocols_render.go (#6424).
At ~680 LOC generateProtocols was the single largest function in the
god-renderer; it owns OSPF/OSPFv3/BGP/RIP/IS-IS stanza emission for the
default instance and every VRF.

Pure code-motion: the function is moved whole and unchanged (same
signature, same body, same output). No logic or render change; the
emitted frr.conf is byte-identical.

Validation: go build ./pkg/frr, gofmt, go vet clean. policy_render.go
drops from 2,202 to 1,522 LOC.
